functions/saveToDB/saveToDB/requirements.txt
google-cloud-pubsub==0.39.1
flask
pymongo